requirements/scorecard.txt
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
PyYAML==6.0.2
requests==2.32.3
